Enable the gitx binary in the CLI
Previously, the gitx module installed the gitx package only. Gitx
ships with a binary that can be symlinked into /usr/local/bin to
enable opening gitx from the command line.  Usually, this involves
enabling the feature in the app itself, but this commit allows the
module to enable the functionality by symlinking the binary into
$boxen::config::bindir.